+++ integration/wayid2nodeids/mapping_test.go
@@ -7,27 +7,6 @@ import (
 	"github.com/Telenav/osrm-backend/integration/nodebasededge"
 )
 
-func TestParseLine(t *testing.T) {
-
-	cases := []struct {
-		line string
-		ids  []int64
-	}{
-		{"", nil},
-		{"24418325,84760891102", nil},
-		{"24418325,84760891102,19496208102", []int64{24418325, 84760891102, 19496208102}},
-		{"24418325,84760891102,19496208102,", []int64{24418325, 84760891102, 19496208102}},
-		{"24418325,84760891102,19496208102,,,,,", []int64{24418325, 84760891102, 19496208102}},
-	}
-
-	for _, c := range cases {
-		result := parseLine(c.line)
-		if !reflect.DeepEqual(result, c.ids) {
-			t.Errorf("parseLine %s, expect %v, but got %v", c.line, c.ids, result)
-		}
-	}
-}
-
 func TestMappingLoad(t *testing.T) {
 
 	m := NewMappingFrom("./testdata/sample_wayid2nodeids.csv.snappy")
